[AGENT-7138] DRUM changes for models()
